% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cvsofa}
\alias{cvsofa}
\title{CV SOFA score calculator}
\usage{
cvsofa(volume = 10, conc = 32, kg = 50)
}
\arguments{
\item{volume}{Volume given. Defaults to 10 mL}

\item{conc}{Concentration of epinephrine or norepinephrine. Defaults to 32 mcg/mL}

\item{kg}{Patient weight in kilograms. Defaults to 50 kg}
}
\description{
Calculate a Sequential Organ Failure Assessment score using volume of norepinephrine, drug concentration, and patient weight in kilograms
}
\examples{
cvsofa()
cvsofa(volume = 11.5, conc = 64, kg = 104)
}
\keyword{CV}
\keyword{SOFA}

