% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mapcalc}
\alias{mapcalc}
\title{Mean Arterial Pressure (MAP) calculator}
\usage{
mapcalc(systolic = 120, diastolic = 80)
}
\arguments{
\item{systolic}{A single or vector of systolic blood pressure(s). Defaults to 120.}

\item{diastolic}{A single or vector of diastolic blood pressure(s). Defaults to 80.}
}
\description{
Mean Arterial Pressure (MAP) calculator
}
\examples{
mapcalc(sysolic = c(120, 100, 90, 114, 89),
        diastolic = c(80, 52,67,98,45))
mapcalc()
}
\keyword{blood}
\keyword{pressure}

