% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{medtools}
\alias{medtools}
\alias{medtools-package}
\title{medtools}
\description{
Medical calculators for use in clinical research.
}
\usage{
cvsofa(volume = 10, conc = 32, kg = 50)
mapcalc(systolic = 120, diastolic = 80)
SF.ratio(o2sat = 90, fio2 = 0.4, liter = 8)
}
\author{Nancy Organ}
