% Generated by roxygen2 (4.0.1): do not edit by hand
\name{SF.ratio}
\alias{SF.ratio}
\title{O2Sat/FiO2 ratio calculator}
\usage{
SF.ratio(o2sat = 90, fio2 = 0.4, liter = 8)
}
\arguments{
\item{o2sat}{Oxygen Saturation. Defaults to 90\%}

\item{fio2}{Fraction of Inspired Oxygen. Defaults to 0.40}

\item{liter}{Liters of Oxygen. Defaults to 8}
}
\description{
Calculate O2Sat/FiO2 ratio from oxygen saturation and Fraction of Inspired Oxygen or Liters of Oxygen. 
}
\examples{
SF.ratio()
[1] 225

SF.ratio(o2sat = 70, liter = 9)
}

